test(db): parallelize migration suite
