+++ test/db_migration_2_test.dart
@@ -1,0 +1,3 @@
+import 'db_migration_suite.dart';
+
+void main() => runDbMigrationTests(1);
