+++ test/db_migration_4_test.dart
@@ -1,0 +1,3 @@
+import 'db_migration_suite.dart';
+
+void main() => runDbMigrationTests(3);
